requirements.txt
altgraph==0.17.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
cffi==1.15.1
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dearpygui==1.9.1
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
fonttools==4.42.0
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
matplotlib==3.7.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
numpy==1.24.4
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pillow-heif==0.13.0
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
Pygments==2.16.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
pyparsing==3.0.9
python-dateutil==2.8.2
pyzmq==25.1.0
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
wcwidth==0.2.6
zipp==3.16.2
